// frontend/src/api/generated/main-service/apiGenerated.ts
export interface ApiKeyDto {
  /** @format uuid */
  id: string
  /** @format uuid */
  userId: string
  /** @format uuid */
  projectId?: null | string
  xApiKey: string
  name?: null | string
  isActive?: null | boolean
  createdBy?: null | string
  /** @format date-time */
  createdDatetime?: any
  updatedBy?: null | string
  /** @format date-time */
  updatedDatetime?: any
}
// ...
export interface UpdateProfileRequest {
  firstName: null | string
  lastName: null | string
  profileImageUrl: null | string
}